requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
huggingface-hub
Pillow
pytesseract
pandas
pdfkit 